+++ tests/templates/test_embeddings/test_displacement_emb.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 """
-Unit tests for the DisplacementEmbedding template.
+Unit tests for the DisplacementEmbedding template.HE
 """
 import pytest
 import numpy as np